refactor: 피드백 반영
- 당첨 번호의 구분자를 view에서 이용하도록 책임 변경
